Updated results fetching to improve performance.
Changed results fetching to map the booksList to BookCard components in the React.useState declaration and matchSorter setResults action. Not completely sure why, but this extremely improves performance. The previous iteration of storing the booksList array in the results state and mapping it to BookCard components in the DisplayBookCards component seemed to cause rerenders on every user input leading to massive performance bottlenecks.