Re-added chip filtering
Re-added chip filtering effect to QuickSearch component. Used the useEffect hook this time to setResults whenever chips changes to avoid needing to setResults in the handleChipClick and handleChipDelete functions. No noticeable difference in performance between the two methods so far, but code is more clean. If performance is an issue, consider updating the filtering algorithm to use the current results list when adding additional filters. I.e., when a new filter is added, the function searches through the whole books list again. However, it would be more efficient to search through the currently filtered list and just apply the one new filter.